# backend/swagger.json
{
    "swagger": "2.0",
    "info": {
    "version": "1.0.0",
    "title": "My User Project CRUD",
    "description": "My User Project Application API",
    "license": {
    "name": "MIT",
    "url": "https://opensource.org/licenses/MIT"
    }
    },
    "host": "localhost:5000",
    "basePath": "/",
    "tags": [
    {
    "name": "Keep my bookmarks",
    "description": "Application to store bookmark links and associated metadata"
    }
    ],
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
        "/bookmarks/getAllEntries": {
        "get": {
        "tags": [
        "Get all"
        ],
        "summary": "Retrieves all entries from the database",
        "parameters": [],
        "responses": {
        "200": {
        "description": "This service returns all records"
        }
        }
        }    
    },
        "/bookmarks/getEntryById": {
            "post": {
            "tags": [
            "Single entry"
            ],
            "summary": "Retrieve single entry from database using ID",
            "parameters": [
            {
            "name": "bmk_id",
            "in": "query",
            "description": "the ID"
            }
            ],
            "responses": {
            "200": {
            "description": "This services returns a single record"
            }
            }
            }    
    },
        "/bookmarks/getEntryByText": {
        "post": {
        "tags": [
        "Multiple entries"
        ],
        "summary": "Retrieve multiple entries from database column title using search string",
        "parameters": [
        {
        "name": "bmk_description",
        "in": "query",
        "description": "search string"
        }
        ],
        "responses": {
        "200": {
        "description": "This services returns all records matching search string",
        "schema": {
            "type": "array" }
        }
        }
        }    
    }     
}
}